Day 19
I'm quite happy with the structure of my code, but slightly less with
the time it takes.

I noticed I'm using quite a lot composed functions. I wonder whether it
would be faster when representating all translations and rotations as
matrices, as this would allow to get rid of these composed functions,
and make a lot of the basic operations (i.e. translating and rotating)
faster.