Day 21 part 2
This was the first time I didn't get to recycle too much code from part
one for part two, and as such it's the first time there are two files
for one day.

I'm quite happy about the speed (less than 50ms on my machine), and that
I finally created a fair number of functions to keep things more or less
comprehensible, although I think there's still more to be done in that
regard.